requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
anyio=3.6.2=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.01.10=haa95532_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=pypi_0
charset-normalizer=3.1.0=pypi_0
colorama=0.4.6=py311haa95532_0
decorator=5.1.1=pyhd3eb1b0_0
executing=0.8.3=pyhd3eb1b0_0
frozenlist=1.3.3=pypi_0
google-api-core=2.11.0=pypi_0
google-api-python-client=2.86.0=pypi_0
google-auth=2.17.3=pypi_0
google-auth-httplib2=0.1.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
googleapis-common-protos=1.59.0=pypi_0
gspread=5.8.0=pypi_0
h11=0.14.0=pypi_0
httpcore=0.16.3=pypi_0
httplib2=0.22.0=pypi_0
httpx=0.23.3=pypi_0
idna=3.4=pypi_0
ipython=8.12.0=py311haa95532_0
jedi=0.18.1=py311haa95532_1
libffi=3.4.2=hd77b12b_6
matplotlib-inline=0.1.6=py311haa95532_0
multidict=6.0.4=pypi_0
oauth2client=4.1.3=pypi_0
oauthlib=3.2.2=pypi_0
openai=0.27.8=pypi_0
openssl=1.1.1t=h2bbff1b_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pip=23.0.1=py311haa95532_0
prompt-toolkit=3.0.36=py311haa95532_0
protobuf=4.22.3=pypi_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
python=3.11.0=h966fe2a_3
python-dotenv=1.0.0=pypi_0
python-telegram-bot=20.2=pypi_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
rfc3986=1.5.0=pypi_0
rsa=4.9=pypi_0
setuptools=66.0.0=py311haa95532_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
tk=8.6.12=h2bbff1b_0
tqdm=4.65.0=pypi_0
traitlets=5.7.1=py311haa95532_0
tzdata=2023c=h04d1e81_0
uritemplate=4.1.1=pypi_0
urllib3=1.26.15=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.38.4=py311haa95532_0
xz=5.2.10=h8cc25b3_1
yarl=1.9.2=pypi_0
zlib=1.2.13=h8cc25b3_0
